rewrite: globalize already set and cache downloads
BREAKING CHANGE

so! basically this is a big big *big* rewrite!! :octocat:

what i've done is introduced a bit of a directory structure where we now save 25-ish images to disk
we choose between those and just pick one

i kinda dislike how we've regressed from the aspect ratio behavior but i'll fix that soon
this rewrite took me *hours*!! so now i'm going to rest for a bit then continue working on it

probably the most interesting feature is that we save a hash of the url to a file such that we do not set the same background twice

although i'm thinking that maybe i did it wrong, let me leave an XXX comment ...

anyways, hope whoever is using this code likes this. (AKA just me)