fix: correct caching handler
this is a big commit, wow but uhh i definitely had to do a lot to fix what I did. the previous version was unimaginabily broken.

Mostly what I did was make it so that the images that're stored are actually downloaded;
What happened before was that the file was created regardless of if the download was succesful.
What we do now is that we create each image in a temporary directory and move it to its final resting place atomically.

What I also did was turn Posts into a stream such that we can keep going after the first page.

I also made it so that the image is resized to fit the screen

Also, I now *actually* remove erroneously downloaded images